Switch to use PHP float type rather than integer for numbers

+++ src/Properties/Factory.php
@@ -26,7 +26,7 @@ class Factory
             }
         }
         if ($attributes->type === 'number') {
-            return new IntegerProperty($name);
+            return new FloatProperty($name);
         } elseif ($attributes->type === 'string' && !isset($attributes->enum)) {
             return new StringProperty($name);
         } elseif ($attributes->type === 'boolean') {
